refactor: move lazy loaded components to search-has-query folder
