+++ src/components/desktop/desktop.vue
@@ -113,9 +113,9 @@
       ScrollToTop,
       SearchBox,
       SlidingPanel,
-      RelatedTags: () => import('../search-has-query.components').then(m => m.RelatedTags),
-      SelectedFilters: () => import('../search-has-query.components').then(m => m.SelectedFilters),
-      Facets: () => import('../search-has-query.components').then(m => m.Facets)
+      RelatedTags: () => import('../search-has-query').then(m => m.RelatedTags),
+      SelectedFilters: () => import('../search-has-query').then(m => m.SelectedFilters),
+      Facets: () => import('../search-has-query').then(m => m.Facets)
     }
   })
   export default class Desktop extends HasQueryMixin {}
