--- src/components/search-has-query.components.ts
@@ -1,23 +1,0 @@
-import {
-  Spellcheck,
-  SpellcheckButton,
-  Redirection,
-  PartialResultsList,
-  PartialQueryButton
-} from '@empathyco/x-components/search';
-import Sort from './sort.vue';
-import Facets from './facets/facets.vue';
-import SelectedFilters from './facets/selected-filters.vue';
-import RelatedTags from './related-tags.vue';
-
-export {
-  Spellcheck,
-  SpellcheckButton,
-  Redirection,
-  PartialResultsList,
-  PartialQueryButton,
-  Sort,
-  Facets,
-  SelectedFilters,
-  RelatedTags
-};
