+++ src/components/search-has-query/facets/number-range-filters.vue
