+++ src/components/search-has-query/facets/price-filter-label.vue
