+++ src/components/search-has-query/facets/searchable-filters.vue
