+++ src/components/search-has-query/facets/selected-filters.vue
