+++ src/components/search-has-query/index.ts
@@ -1,0 +1,18 @@
+import { PartialResultsList, PartialQueryButton } from '@empathyco/x-components/search';
+import Facets from './facets/facets.vue';
+import SelectedFilters from './facets/selected-filters.vue';
+import RelatedTags from './related-tags.vue';
+import SpellcheckMessage from './spellcheck-message.vue';
+import NoResultsMessage from './no-results-message.vue';
+import Redirection from './redirection.vue';
+
+export {
+  Redirection,
+  PartialResultsList,
+  PartialQueryButton,
+  Facets,
+  SelectedFilters,
+  RelatedTags,
+  SpellcheckMessage,
+  NoResultsMessage
+};
